+++ NEWS.md
@@ -6,8 +6,8 @@
 * New `ps_wait()` to start an interruptible wait on multiple processes,
   with a timeout (#166).
 
-* `ps_handle()` now allowes a non-integer object as the pid, as long as
-  its value is integer.
+* `ps_handle()` now allowes a numeric (double) scalar as the pid, as long
+  as its value is integer.
 
 * `ps_kill()` can now kill multiple processes concurrently.